+++ refena/lib/refena.dart
@@ -12,6 +12,8 @@ export 'package:refena/src/async_value_join.dart';
 export 'package:refena/src/container.dart';
 export 'package:refena/src/notifier/base_notifier.dart'
     show
+        NotifyStrategy,
+        ViewProviderNotifier,
         ReduxNotifierOverrideExt,
         GlobalReduxNotifierOverrideExt,
         MockReducer,
@@ -19,8 +21,6 @@ export 'package:refena/src/notifier/base_notifier.dart'
         NotifierTester,
         AsyncNotifierTester,
         ReduxNotifierTester;
-export 'package:refena/src/notifier/base_notifier.dart'
-    show NotifyStrategy, ViewProviderNotifier;
 export 'package:refena/src/notifier/notifier_event.dart';
 export 'package:refena/src/notifier/types/async_notifier.dart';
 export 'package:refena/src/notifier/types/change_notifier.dart';
@@ -32,6 +32,7 @@ export 'package:refena/src/notifier/types/pure_notifier.dart';
 export 'package:refena/src/notifier/types/redux_notifier.dart';
 export 'package:refena/src/notifier/types/state_notifier.dart';
 export 'package:refena/src/notifier/types/stream_provider_notifier.dart';
+export 'package:refena/src/notifier/types/view_family_provider_notifier.dart';
 export 'package:refena/src/observer/error_parser.dart' show ErrorParser;
 export 'package:refena/src/observer/event.dart';
 export 'package:refena/src/observer/history_observer.dart';
@@ -47,6 +48,7 @@ export 'package:refena/src/provider/types/provider.dart';
 export 'package:refena/src/provider/types/redux_provider.dart';
 export 'package:refena/src/provider/types/state_provider.dart';
 export 'package:refena/src/provider/types/stream_provider.dart';
+export 'package:refena/src/provider/types/view_family_provider.dart';
 export 'package:refena/src/provider/types/view_provider.dart';
 export 'package:refena/src/ref.dart' show Ref, WatchableRef;
 export 'package:refena/src/reference.dart';