+++ refena_flutter/lib/refena_flutter.dart
@@ -6,4 +6,5 @@ export 'package:refena_flutter/src/mixin.dart';
 export 'package:refena_flutter/src/scope.dart' show RefenaScope;
 export 'package:refena_flutter/src/tracing/tracing_page.dart';
 export 'package:refena_flutter/src/view_model_builder.dart';
+export 'package:refena_flutter/src/view_model_param_builder.dart';
 export 'package:refena_flutter/src/widget_rebuildable.dart';