+++ refena/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.45.0
+
+- feat: add `ViewFamilyProvider` and `ViewModelParamBuilder` as a replacement for `provider.family` of Riverpod
+- feat: `ref.read` should support family providers and `select`
+
 ## 0.44.0
 
 - feat: add `container.exists`, `container.getActiveProviders`